fix(storage): genesis receipts should be saved
